flambda2-types: Do not expose Typing_env_extension.meet
We are only using env extension meet in a single place outside of the
typing env (`simplify_variadic_primitive`), and there is no good reason
to do an env extension meet there since we are going to add the
resulting meet to the typing env anyways.

This simplifies the API surface.