Improved meet for naked immediates
This is one part of #4112, with information stored in the types in the
style of #2259 (which it replaces).

One crucial difference with #2259 is that during the meet, we
only reduce new information. For instance, if we are meeting:

  (Naked_immediate {0, 1, 2} (Get_tag x))

and

  (Naked_immediate {1, 2} (Get_tag y))

then we will propagate:

 - Nothing if `x` and `y` have the same canonical;

 - The set of tags `{1, 2}` on `x` otherwise.

This provides a way to solve potential reduction loops with forward
reductions from the value itself: if the set of possible tags of `x`
shrinks to a new set `S`, and we know that `get_tag(x) = get_tag_x`,
then we can add the equation

	get_tag_x : (Naked_immediate S (Get_tag x))

If we already know this set of possible values for `get_tag_x`, then
nothing happens. If we know (or learn through intersection with `S`) a
more precise set of values for `get_tag_x`, the reduction on `get_tag_x`
will add a type on `x` with tags `S'` and `get_tag` variable
`get_tag_x`. Since `S'` is a subset of `S`, we will then not propagate
things back onto `get_tag_x`.